app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
redis==5.0.7
psycopg2-binary==2.9.9
meilisearch==0.32.0
qdrant-client==1.9.1
numpy==1.26.4
scikit-learn==1.5.1
onnxruntime==1.18.0
sentence-transformers==3.0.1
transformers==4.43.3
optimum==1.21.4
tokenizers==0.19.1
python-dotenv==1.0.1
kiwipiepy==0.21.0
